requirements.txt
requests
python-dotenv
PyQt5